Fixed merge related conflicts
